Optimised data extraction, Z and W
Changed the way the data is extracted from the file and used within the main for better optimisation/future use.

Added in a plot of the Z boson pt.

Added my preliminary calculation of the W boson four vector, something has gone wrong as the relativistic energy is below the rest mass.